Give Staff one Client URL that branches on kind so Players and Coaches routes can go. (#40)
