+++ apps/dashboard/src/actions/trash/trashClient.ts
@@ -15,8 +15,7 @@ export const trashClientAction = staffActionClient
 
     revalidatePath("/clients");
     revalidatePath("/trash");
-    revalidatePath(`/players/${parsedInput.id}`);
-    revalidatePath(`/coaches/${parsedInput.id}`);
+    revalidatePath(`/clients/${parsedInput.id}`);
 
     return { ok: true as const };
   });
